+++ app/src/features/contentProduction/index.ts
@@ -1,0 +1,1 @@
+export { ContentProductionPage } from './pages/ContentProductionPage'